flask-server/requirements.txt
requests
flask
python-dotenv
gunicorn
